+++ modules/core-config/modules/external-dns/identity.tf
@@ -5,7 +5,11 @@ module "identity_private" {
   location            = var.location
   resource_group_name = var.resource_group_name
 
+  workload_identity = var.workload_identity && local.use_aad_workload_identity
+  oidc_issuer_url   = var.cluster_oidc_issuer_url
+
   name      = "${var.cluster_name}-external-dns-private"
+  subjects  = ["system:serviceaccount:${var.namespace}:${local.private_service_account_name}"]
   namespace = var.namespace
   labels    = var.labels
 
@@ -27,7 +31,11 @@ module "identity_public" {
   location            = var.location
   resource_group_name = var.resource_group_name
 
+  workload_identity = var.workload_identity && local.use_aad_workload_identity
+  oidc_issuer_url   = var.cluster_oidc_issuer_url
+
   name      = "${var.cluster_name}-external-dns-public"
+  subjects  = ["system:serviceaccount:${var.namespace}:${local.public_service_account_name}"]
   namespace = var.namespace
   labels    = var.labels