+++ modules/core-config/modules/fluentd/identity.tf
@@ -4,7 +4,11 @@ module "identity" {
   location            = var.location
   resource_group_name = var.resource_group_name
 
+  workload_identity = var.workload_identity && local.use_aad_workload_identity
+  oidc_issuer_url   = var.cluster_oidc_issuer_url
+
   name      = "${var.cluster_name}-fluentd"
+  subjects  = ["system:serviceaccount:${var.namespace}:${local.service_account_name}"]
   namespace = var.namespace
   labels    = var.labels