+++ variables.tf
@@ -425,6 +425,7 @@ variable "experimental" {
     fluent_bit_use_memory_buffer                               = optional(bool, false)
     fluentd_memory_override                                    = optional(string, null)
     prometheus_memory_override                                 = optional(string, null)
+    workload_identity                                          = optional(bool, false)
   })
   nullable = false
   default  = {}